Refactor tools cross-referencing each other, plus cleanup:
- Remove AssetView class which solely set 3 properties on the TreeView
- Refactor `Node` to `Item` as it is for QAbstractItemModels
- Refactor some variables to `snake_case` as opposed to `lowerCamelCase`
- Add some todo notes (notes to self for cleanup)
