+++ avalon/tools/models.py
@@ -1,0 +1,421 @@
+import re
+import logging
+import collections
+
+from ..vendor.Qt import QtCore, QtWidgets
+from ..vendor import qtawesome as awesome
+from .. import io
+from .. import style
+
+log = logging.getLogger(__name__)
+
+
+class TreeModel(QtCore.QAbstractItemModel):
+
+    COLUMNS = list()
+    ItemRole = QtCore.Qt.UserRole + 1
+
+    def __init__(self, parent=None):
+        super(TreeModel, self).__init__(parent)
+        self._root_item = Item()
+
+    def rowCount(self, parent):
+        if parent.isValid():
+            item = parent.internalPointer()
+        else:
+            item = self._root_item
+
+        return item.childCount()
+
+    def columnCount(self, parent):
+        return len(self.COLUMNS)
+
+    def data(self, index, role):
+
+        if not index.isValid():
+            return None
+
+        if role == QtCore.Qt.DisplayRole or role == QtCore.Qt.EditRole:
+
+            item = index.internalPointer()
+            column = index.column()
+
+            key = self.COLUMNS[column]
+            return item.get(key, None)
+
+        if role == self.ItemRole:
+            return index.internalPointer()
+
+    def setData(self, index, value, role=QtCore.Qt.EditRole):
+        """Change the data on the nodes.
+
+        Returns:
+            bool: Whether the edit was successful
+        """
+
+        if index.isValid():
+            if role == QtCore.Qt.EditRole:
+
+                item = index.internalPointer()
+                column = index.column()
+                key = self.COLUMNS[column]
+                item[key] = value
+
+                # passing `list()` for PyQt5 (see PYSIDE-462)
+                self.dataChanged.emit(index, index, list())
+
+                # must return true if successful
+                return True
+
+        return False
+
+    def setColumns(self, keys):
+        assert isinstance(keys, (list, tuple))
+        self.COLUMNS = keys
+
+    def headerData(self, section, orientation, role):
+
+        if role == QtCore.Qt.DisplayRole:
+            if section < len(self.COLUMNS):
+                return self.COLUMNS[section]
+
+        super(TreeModel, self).headerData(section, orientation, role)
+
+    def flags(self, index):
+        flags = QtCore.Qt.ItemIsEnabled
+
+        item = index.internalPointer()
+        if item.get("enabled", True):
+            flags |= QtCore.Qt.ItemIsSelectable
+
+        return flags
+
+    def parent(self, index):
+
+        item = index.internalPointer()
+        parent_node = item.parent()
+
+        # If it has no parents we return invalid
+        if parent_node == self._root_item or not parent_node:
+            return QtCore.QModelIndex()
+
+        return self.createIndex(parent_node.row(), 0, parent_node)
+
+    def index(self, row, column, parent):
+        """Return index for row/column under parent"""
+
+        if not parent.isValid():
+            parent_item = self._root_item
+        else:
+            parent_item = parent.internalPointer()
+
+        child_item = parent_item.child(row)
+        if child_item:
+            return self.createIndex(row, column, child_item)
+        else:
+            return QtCore.QModelIndex()
+
+    def add_child(self, node, parent=None):
+        if parent is None:
+            parent = self._root_item
+
+        parent.add_child(node)
+
+    def column_name(self, column):
+        """Return column key by index"""
+
+        if column < len(self.COLUMNS):
+            return self.COLUMNS[column]
+
+    def clear(self):
+        self.beginResetModel()
+        self._root_item = Item()
+        self.endResetModel()
+
+
+class Item(dict):
+    """An item that can be represented in a tree view using `TreeModel`.
+
+    The item can store data just like a regular dictionary.
+
+    >>> data = {"name": "John", "score": 10}
+    >>> item = Item(data)
+    >>> assert item["name"] == "John"
+
+    """
+
+    def __init__(self, data=None):
+        super(Item, self).__init__()
+
+        self._children = list()
+        self._parent = None
+
+        if data is not None:
+            assert isinstance(data, dict)
+            self.update(data)
+
+    def childCount(self):
+        return len(self._children)
+
+    def child(self, row):
+
+        if row >= len(self._children):
+            log.warning("Invalid row as child: {0}".format(row))
+            return
+
+        return self._children[row]
+
+    def children(self):
+        return self._children
+
+    def parent(self):
+        return self._parent
+
+    def row(self):
+        """
+        Returns:
+             int: Index of this node under parent"""
+        if self._parent is not None:
+            siblings = self.parent().children()
+            return siblings.index(self)
+
+    def add_child(self, child):
+        """Add a child to this node"""
+        child._parent = self
+        self._children.append(child)
+
+
+class TasksModel(TreeModel):
+    """A model listing the tasks combined for a list of assets"""
+
+    COLUMNS = ["name", "count"]
+
+    def __init__(self):
+        super(TasksModel, self).__init__()
+        self._num_assets = 0
+        self._icons = {
+            "__default__": awesome.icon("fa.male", color=style.colors.default)
+        }
+
+        self._get_task_icons()
+
+    def _get_task_icons(self):
+        # Get the project configured icons from database
+        project = io.find_one({"type": "project"})
+        tasks = project['config'].get('tasks', [])
+        for task in tasks:
+            icon_name = task.get("icon", None)
+            if icon_name:
+                icon = awesome.icon("fa.{}".format(icon_name),
+                                    color=style.colors.default)
+                self._icons[task["name"]] = icon
+
+    def set_assets(self, asset_ids):
+        """Set assets to track by their database id
+
+        Arguments:
+            asset_ids (list): List of asset ids.
+
+        """
+
+        assets = list()
+        for asset_id in asset_ids:
+            asset = io.find_one({"_id": asset_id, "type": "asset"})
+            assert asset, "Asset not found by id: {0}".format(asset_id)
+            assets.append(asset)
+
+        self._num_assets = len(assets)
+
+        tasks = collections.Counter()
+        for asset in assets:
+            asset_tasks = asset.get("data", {}).get("tasks", [])
+            tasks.update(asset_tasks)
+
+        self.clear()
+        self.beginResetModel()
+
+        default_icon = self._icons["__default__"]
+
+        if not tasks:
+            node = Item({
+                "name": "No task",
+                "count": 0,
+                "icon": default_icon,
+                "enabled": False,
+            })
+
+            self.add_child(node)
+
+        else:
+            for task, count in sorted(tasks.items()):
+                icon = self._icons.get(task, default_icon)
+
+                node = Item({
+                    "name": task,
+                    "count": count,
+                    "icon": icon
+                })
+
+                self.add_child(node)
+
+        self.endResetModel()
+
+    def headerData(self, section, orientation, role):
+
+        # Override header for count column to show amount of assets
+        # it is listing the tasks for
+        if role == QtCore.Qt.DisplayRole:
+            if orientation == QtCore.Qt.Horizontal:
+                if section == 1:  # count column
+                    return "count ({0})".format(self._num_assets)
+
+        return super(TasksModel, self).headerData(section, orientation, role)
+
+    def data(self, index, role):
+
+        if not index.isValid():
+            return
+
+        # Add icon to the first column
+        if role == QtCore.Qt.DecorationRole:
+            if index.column() == 0:
+                return index.internalPointer()['icon']
+
+        return super(TasksModel, self).data(index, role)
+
+
+class AssetModel(TreeModel):
+    """A model listing assets in the silo in the active project.
+
+    The assets are displayed in a treeview, they are visually parented by
+    a `visualParent` field in the database containing an `_id` to a parent
+    asset.
+
+    """
+
+    COLUMNS = ["label"]
+    Name = 0
+    Deprecated = 2
+    ObjectId = 3
+
+    DocumentRole = QtCore.Qt.UserRole + 2
+    ObjectIdRole = QtCore.Qt.UserRole + 3
+
+    def __init__(self, silo=None, parent=None):
+        super(AssetModel, self).__init__(parent=parent)
+
+        self._silo = None
+
+        if silo is not None:
+            self.set_silo(silo, refresh=True)
+
+    def set_silo(self, silo, refresh=True):
+        """Set the root path to the ItemType root."""
+        self._silo = silo
+        if refresh:
+            self.refresh()
+
+    def _add_hierarchy(self, parent=None):
+
+        # Find the assets under the parent
+        find_data = {
+            "type": "asset",
+            "silo": self._silo,
+        }
+        if parent is None:
+            # if not a parent find all that are parented to the project
+            # or do *not* have a visualParent field at all
+            find_data['$or'] = [
+                {'data.visualParent': {'$exists': False}},
+                {'data.visualParent': None}
+            ]
+        else:
+            find_data["data.visualParent"] = parent['_id']
+
+        assets = io.find(find_data).sort('name', 1)
+
+        for asset in assets:
+
+            # get label from data, otherwise use name
+            data = asset.get("data", {})
+            label = data.get("label", asset['name'])
+            tags = data.get("tags", [])
+
+            # store for the asset for optimization
+            deprecated = "deprecated" in tags
+
+            node = Item({
+                "_id": asset['_id'],
+                "name": asset["name"],
+                "label": label,
+                "type": asset['type'],
+                "tags": ", ".join(tags),
+                "deprecated": deprecated,
+                "_document": asset
+            })
+            self.add_child(node, parent=parent)
+
+            # Add asset's children recursively
+            self._add_hierarchy(node)
+
+    def refresh(self):
+        """Refresh the data for the model."""
+
+        self.clear()
+        self.beginResetModel()
+        if self._silo:
+            self._add_hierarchy(parent=None)
+        self.endResetModel()
+
+    def flags(self, index):
+        return QtCore.Qt.ItemIsEnabled | QtCore.Qt.ItemIsSelectable
+
+    def data(self, index, role):
+
+        if not index.isValid():
+            return
+
+        node = index.internalPointer()
+        if role == QtCore.Qt.DecorationRole:        # icon
+
+            column = index.column()
+            if column == self.Name:
+
+                # Allow a custom icon and custom icon color to be defined
+                data = node["_document"]["data"]
+                icon = data.get("icon", None)
+                color = data.get("color", style.colors.default)
+
+                if icon is None:
+                    # Use default icons if no custom one is specified.
+                    # If it has children show a full folder, otherwise
+                    # show an open folder
+                    has_children = self.rowCount(index) > 0
+                    icon = "folder" if has_children else "folder-o"
+
+                # Make the color darker when the asset is deprecated
+                if node.get("deprecated", False):
+                    color = QtGui.QColor(color).darker(250)
+
+                try:
+                    key = "fa.{0}".format(icon)  # font-awesome key
+                    icon = awesome.icon(key, color=color)
+                    return icon
+                except Exception as exception:
+                    # Log an error message instead of erroring out completely
+                    # when the icon couldn't be created (e.g. invalid name)
+                    log.error(exception)
+
+                return
+
+        if role == QtCore.Qt.ForegroundRole:        # font color
+            if "deprecated" in node.get("tags", []):
+                return QtGui.QColor(style.colors.light).darker(250)
+
+        if role == self.ObjectIdRole:
+            return node.get("_id", None)
+
+        if role == self.DocumentRole:
+            return node.get("_document", None)
+
+        return super(AssetModel, self).data(index, role)
